Rename ProofDeleteChip to ProofDeleteConfirmationDialog
